### 2023-06-12 #5546 @editorialbot
[PRE REVIEW]: SMART: Spatial Modeling Algorithms for Reactions and Transport
**Submitting author:** @emmetfrancis (<a href="http://orcid.org/0000-0001-6070-805X">Emmet A. Francis</a>)
**Repository:** https://github.com/RangamaniLabUCSD/smart
**Branch with paper.md** (empty if default branch):
**Version:** v2.1.1
