### 2023-06-17 #5546 @Kevin-Mattheus-Moerman
:wave: @MathCancer @mbarzegary would any of you be willing to review this submission for [JOSS](https://joss.theoj.org/about)? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.05546/joss.05546/10.21105.joss.05546.pdf
Software: https://github.com/RangamaniLabUCSD/smart
