### 2023-06-22 ff0fa25
Update implementation.md
Adding a diagram illustrating the SMART workflow, along with some brief text, on the "software implementation" page of the docs.
