### 2023-06-26 justinlaughlin/stubs#118 @mbarzegary
[JOSS Review] Installation
Dear authors,

The software installation routine is straightforward, and there is no special issue with it. However, I found the package a bit difficult to get started with, especially for a normal user. If the package is supposed to be used by advanced users only (which I think is not), it should be explicitly mentioned.

* After the installation using the docker approach, how is the user supposed to work with the examples? I think there is no way in the current setup. Jupyter is not installed in the docker image, and there is no port forwarding on the docker run command. I would say a non-advanced user can never find out how to do that, so it can be beneficial to guide the users on how to run the example codes and see how the library works in action. I did it by running `jupyter notebook --ip 0.0.0.0 --port 8888 --no-browser --allow-root` inside the container (after forwarding the port in the docker execution command), in which the inline PyVista visualization didn't still work (which is because PyVista is not installed inside the container).
* As a follow-up to the above point, it can be nice if the user can have access to the example codes in separate files without the Jupyter notebooks, similar to example 3. These codes (accompanied by some documentation on how to run them) can show a normal user how to use the package after the installation.

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/5580
