### 2023-06-26 justinlaughlin/stubs#119 @mbarzegary
 [JOSS Review] Reproducibility and functionality
Dear authors,

Generally speaking, the software runs as described in the documentation. However, there are very minor comments in this regard:

* The software claims to be high-performance, but no performance test is reported. Having a small test on example 3 (as the relevant example on parallelization) can be nice in this regard.
* Is there any way to reproduce Figure 1 of the paper? Example 2 performs the simulation on the mentioned model on a simple geometry.

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/5580
