### 2023-06-26 justinlaughlin/stubs#120 @mbarzegary
[JOSS Review] Tests
Dear authors,

The software has standard and comprehensive automated tests implemented using GitHub workflows. However, it's not very clear how a user/developer can test a local installation of the code. Inside the docker container, running `pytest` (or `python3 -m pytest`) command returns this error:
```
_main__.py: error: unrecognized arguments: --cov=smart --cov-report --cov-report term-missing
  inifile: /home/shared/pytest.ini
  rootdir: /home/shared
```

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/5580
