### 2023-06-26 justinlaughlin/stubs#121 @mbarzegary
[JOSS Review] Documentation 
Dear authors,

The documentation of the software is nicely done and contains an overview of the underlying mathematical and computational aspects too. However, you may consider improving it with these suggestions:

* It can be beneficial to talk a little bit about the visualization of results in an external postprocessor like ParaView. For all the examples, the results are written on the disc, but the documentation and examples don’t mention how to deal with them except by using the provided analysis tools. If you assume that the user should already know about how FEniCS and its IO pipelines work, you need to mention these prerequisites explicitly (also in the paper).
* For the examples, it can be beneficial to mention where the values of the model parameters are coming from. This can help the users build their own models from scratch. For example, are the chosen values of parameters in example 6 based on the model of Doi et al.? If not, how are the values chosen?
* It can be beneficial to mention that some of the cells of the Jupyter files take longer to compute, maybe with an approximation of the time that the user needs to wait for the results to come. For instance, the loop in example 3 seems to never end (it took half an hour to finish for me I think). The same problem exists with example 6.
* How is an advanced user supposed to change the performance-related configs, like the default PETSc preconditioner and solver types and parameters used by SMART? Is that a possibility via the `configCur.solver.update` function?

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/5580
