### 2023-06-26 #5580 @mbarzegary
I confirm the quality of the software and the effort the author has put in to develop it. The developed package is successful in reproducing the previously published simulations using its simple, yet powerful data structure. Moreover, the software includes nice documentation and a bunch of tutorials covering most of the features of the package. However, I believe some minor concerns should be addressed before accepting the submission, elaborated in a couple of issues I have opened on the software repo (listed above). I hope the authors find the comments useful.
