### 2023-06-27 e7d8fbd
Add image including pyvista and panel for plotting with jupyterlab
