### 2023-06-27 0b14d2e
Remove inline from various examples and fix readmes
