+++ .github/workflows/docker-image.yml
@@ -41,7 +41,7 @@ jobs:
           username: ${{ github.actor }}
           password: ${{ secrets.GITHUB_TOKEN }}
 
-      - name: Extract metadata (tags, labels) for base-iamge
+      - name: Extract metadata (tags, labels) for base-image
         id: meta-base
         uses: docker/metadata-action@v4
         with:
