### 2023-06-27 9badb63
Add pytest-cov to resolve https://github.com/RangamaniLabUCSD/smart/issues/120
