+++ pyproject.toml
@@ -43,17 +43,21 @@ docs = [
 examples = [
     "meshio",
     "gmsh",
-    "pyvista==0.38.4",
-    "panel",
     "matplotlib",
     "jupyter",
     "jupyterlab"
 ]
+pyvista = [
+    "pyvista==0.38.4",
+    "panel"
+]
+
 all = [
    "smart[examples]",
    "smart[test]",
    "smart[docs]",
-   "smart[dev]"
+   "smart[dev]",
+   "smart[pyvista]"
 ]
 
 
