+++ Dockerfile
@@ -11,8 +11,7 @@ WORKDIR /repo
 RUN python3 -m pip install ".[test,examples]"
 RUN dpkgArch="$(dpkg --print-architecture)"; \
     case "$dpkgArch" in amd64) \
-    python3 -m pip install ".[pyvista]" \
-    python3 -m pip cache purge ;; \
+    python3 -m pip install ".[pyvista]" ;; \
     esac;
 
 RUN python3 -m pip install pre-commit
