### 2023-06-27 justinlaughlin/stubs#118 @jorgensd
@mbarzegary I've tried to address the comments above in the #123.

Some notes:
- The PR changes the base image to have `jupyter`, `jupyterlab` and other dependencies
  - On linux with ARM build, VTK does not ship binaries, so we cannot build `pyvista` on that platform (as the build time exceeds the Github CI runtime). Thus only `linux/amd64` images will have pyvista.
- A separate image will be pushed to the Github Registries, which uses `jupyter lab` as entrypoint, with the necessary arguments.
- The source code of smart is now located in `/repo` in the container, which is the default working directory if not overloaded by a user at container initialization.
