### 2023-06-28 12ca69c
Black formatting of visualization + skip pyvista if not installed
