### 2023-06-28 5db3a0a
Add pyvista to docker image
