### 2023-06-29 204fd93
Add 'getting started' to the docs and add some fixes for consistency throughout the examples
