### 2023-06-30 c8b4f41
Update docs/getting_started.md
Co-authored-by: Jørgen Schartum Dokken <dokken92@gmail.com>
