### 2023-06-30 bfd5502
Make suggested changes to getting_started.md, make current solution accessible at species.sol (single line change in model.py)
