+++ docs/getting_started.md
@@ -17,26 +17,27 @@ D_unit = smart.units.unit.um**2 / smart.units.unit.s
 conc_unit = smart.units.unit.molecule / smart.units.unit.um**2
 ```
 Following these preliminary steps, we begin the basic workflow of SMART:
-1. **Initialize SMART containers**: Each container (species, compartments, reactions, or parameters) is initialized by first creating individual objects of the class `smart.model_assembly.Species`, `smart.model_assembly.Compartment`, `smart.model_assembly.Reaction`, `smart.model_assembly.Parameter`. The initialization of these objects is described in some detail in Examples 1 and 3, or comprehensively in the API documentation.
+1. **Initialize SMART containers**: Each container (species, compartments, reactions, or parameters) is initialized by first creating individual objects of the class `smart.model_assembly.Species`, `smart.model_assembly.Compartment`, `smart.model_assembly.Reaction`, `smart.model_assembly.Parameter`. The initialization of these objects is described in some detail in Examples [1](https://rangamanilabucsd.github.io/smart/examples/example1/example1.html) and [3](https://rangamanilabucsd.github.io/smart/examples/example3/example3.html), or more comprehensively in the [API documentation](https://rangamanilabucsd.github.io/smart/autoapi/smart/model_assembly/index.html#module-smart.model_assembly).
 
-    These SMART containers contain all instances of any given class. In general, the user should define an empty container of the correct type and then add all instances of each object manually. In this simple example, we consider a model with a single species (`A`) over a single surface (`surf`), with a single reaction (`r1`, with parameter `k` defining the rate of decay) corresponding to exponential decay over time:
+    These SMART containers contain all instances of any given class. In general, the user should define an empty container of the correct type and then add all instances of each object manually. In this simple example, we consider a model with a single species (`A`) over a single surface (`surf`), with a single reaction (`r1`, with parameter `k` defining the rate of decay) corresponding to exponential decay over time. Note that the `Reaction` constructor takes in an equation written as a string, whose input arguments, `k` and `A`, are mapped to either a `Species` or `Parameter`. The input argument in the `param_map` maps from the equation variable name (`"k"`) to the name of the parameter (`"kdecay"`). Similarly, the input argument in the `species_map` maps from the equation variable name (`"sp"`) to the name of the species (`"A"`).
     ```python
-    surf = smart.model_assembly.Compartment("surf", 2, smart.units.unit.um, 1)
-    A = smart.model_assembly.Species("A", 10, conc_unit, 1.0, D_unit, "surf")
-    k = smart.model_assembly.Parameter("k", 1, 1/smart.units.unit.s)
-    r1 = smart.model_assembly.Reaction("r1", ["A"], [], param_map={"k":"k"}, eqn_f_str="k*A")
+    surf_var = smart.model_assembly.Compartment("surf", 2, smart.units.unit.um, 1)
+    A_var = smart.model_assembly.Species("A", 10, conc_unit, 1.0, D_unit, "surf")
+    kdecay_var = smart.model_assembly.Parameter("kdecay", 1, 1/smart.units.unit.s)
+    r1_var = smart.model_assembly.Reaction("r1", ["A"], [], param_map={"k":"kdecay"},
+                                            species_map={"sp":"A"}, eqn_f_str="k*sp")
     cc = smart.model_assembly.CompartmentContainer()
-    cc.add([surf])
+    cc.add([surf_var])
     sc = smart.model_assembly.SpeciesContainer()
-    sc.add([A])
[…]
