### 2023-07-04 8e11a75
Add suggested property to species class
