### 2023-07-04 963707c
Add 3d version of Example 2 with realistic dendritic spine geometry
