### 2023-07-04 #5580 @oalii
Dear Authors,

First, sorry for the detail of my response.
Secondly, as @mbarzegary  I acknowledge the overall quality of your code and documentation ! You put a lot of efforts into it and it is really appreciated.
But I struggled with the docker installation, I left an issue with a bit more details.
Some potential users (me included) are more familiar with conda environments than docker containers. Would it be possible to provide a conda recipe to install and run you library (This could widen the spectrum of potential users) ?  This is just a question, not a request, as I suspect that the available Fenics version for conda might be incompatible.
As soon as the docker issue will be fixed, I'll dive into your notebooks.
