+++ docs/getting_started.md
@@ -4,7 +4,7 @@ The basic SMART workflow is summarized in the diagram below. This is the general
 
 ![SMART flow chart](images/SMART_flow_chart.png)
 
-To use SMART, the user should plan to write a python script (or Jupyter notebook) that goes through the outlined steps. The easiest way to get started with this is to follow the template of the SMART demo Jupyter notebooks or .py files. In the `smart` and `smart-lab` images, these files exist in the `/repo/examples` folder.
+To use SMART, the user should plan to write a Python script (or Jupyter notebook) that goes through the outlined steps. The easiest way to get started with this is to follow the template of the SMART demo Jupyter notebooks (`.ipynb`) or `.py` files. In the `smart` and `smart-lab` images, these files exist in the `/repo/examples` folder.
 
 Here, we walk through a minimal example demonstrating the use of SMART. In this simple example, the only required imports are `smart` and `dolfin`:
 ```python
@@ -35,7 +35,7 @@ Following these preliminary steps, we begin the basic workflow of SMART:
     rc = smart.model_assembly.ReactionContainer()
     rc.add([r1_var])
     ```
-2. **Generate/load in mesh**: Separately from defining the SMART containers, the user must either generate or load in a mesh. In Examples 2, 3, 4, and 6, this mesh is generated using [Gmsh](https://gmsh.info/), via our module `smart.mesh_tools`. Then [meshio](https://github.com/nschloe/meshio) is used to convert the .msh file generated by [Gmsh](https://gmsh.info/) into a mesh for use by DOLFIN (the computational backend of [FEniCS](https://fenicsproject.org/)). Alternatively, Examples 1 and 5 use built-in mesh functions in DOLFIN (either `dolfin.RectangleMesh()` or `dolfin.UnitCubeMesh()`). For some applications, the user may wish to generate a mesh using other software. These can then be loaded in for use by SMART by using [meshio](https://github.com/nschloe/meshio) to convert to an [XDMF-file](https://www.xdmf.org/index.php/XDMF_Model_and_Format) and loading into dolfin using `dolfin.XDMFFile()`. (see `smart.mesh_tools.gmsh_to_dolfin` for an example).
+2. **Generate/load in mesh**: Separately from defining the SMART containers, the user must either generate or load in a mesh. In Examples 2, 3, 4, and 6, this mesh is generated using [Gmsh](https://gmsh.info/), via our module `smart.mesh_tools`. Then [meshio](https://github.com/nschloe/meshio) is used to convert the `.msh` file generated by [Gmsh](https://gmsh.info/) into a mesh for use by DOLFIN (the computational backend of [FEniCS](https://fenicsproject.org/)). Alternatively, Examples 1 and 5 use built-in mesh functions in DOLFIN (either `dolfin.RectangleMesh()` or `dolfin.UnitCubeMesh()`). For some applications, the user may wish to generate a mesh using other software. These can then be loaded in for use by SMART by using [meshio](https://github.com/nschloe/meshio) to convert to an [XDMF-file](https://www.xdmf.org/index.php/XDMF_Model_and_Format) and loading into dolfin using `dolfin.XDMFFile()`. (see `smart.mesh_tools.gmsh_to_dolfin` for an example).
 
[…]
