### 2023-07-05 92d69c3
Add 3D example to documentation
