+++ .github/workflows/build_docs.yml
@@ -40,7 +40,7 @@ jobs:
         uses: actions/cache@v3
         with:
           path: _build
-          key: cache_key_2  # Bump nuber manually when you want to clear the cache
+          key: cache_key_2  # Bump number manually when you want to clear the cache
 
       - name: Build docs
         run: jupyter book build .
