### 2023-07-06 d51fca1
Add clip_logic and clim options to visualization functions
