### 2023-07-06 edc9124
Remove instances of print_assembly from examples, model.py, and solvers.py
