### 2023-07-07 82d701e
Return image from plot when saving
