### 2023-07-07 78a10f9
Try PYVISTA_OFF_SCREEN=True and return image directly from the plotting object
