+++ .github/workflows/build_docs.yml
@@ -22,7 +22,7 @@ jobs:
       DISPLAY: ":99.0"
       PYVISTA_TRAME_SERVER_PROXY_PREFIX: '/proxy/'
       PYVISTA_TRAME_SERVER_PROXY_ENABLED: "True"
-      PYVISTA_OFF_SCREEN: false
+      PYVISTA_OFF_SCREEN: True
       PYVISTA_JUPYTER_BACKEND: "panel"
 
     steps:
