### 2023-07-07 142f9df
Add pyvista dependencies to build_docs
