### 2023-07-07 justinlaughlin/stubs#119 @emmetfrancis
We now have introduced a version of Example 2 in 3D (as shown in Figure 1), which is included in our online documentation as well. We are in the process of considering possible performance tests that could be added.
