### 2023-07-11 #5580 @emmetfrancis
Thank you for checking in, we have resolved almost all of the issues posted by @oalii and @mbarzegary. The only issue we have not addressed is the suggestion by @mbarzegary to add performance testing. The suggestion to add a performance test associated with parallelization is a good one, but we are still working to add some parallelization features in future versions of SMART. We will follow up on this issue soon, we appreciate all the feedback and suggestions!
