### 2023-07-12 #5580 @vincent-noel
Dear authors,

Congratulations on the great work !
The software is very well done, well documented, tested, and with many examples. Running the examples via the smart-lab docker image is extremely easy and provide a ready-to-use environment to start experimenting right away.
As @oalii said, conda compatibility would be best, but these things can take time and I'm sure they will be resolved. Meanwhile providing everything in a docker environment works very well.
I also agree with @mbarzegary that even a basic performance test would be a very good addition. I would also like to see a mention of what kind of parallelization SMART supports, which is something users would expect from a "high performance" software.
