### 2023-07-12 #5580 @jorgensd
SMART supports parallelism through MPI.
A note on the conda/DOLFIN status:
I've made three PRs into DOLFIN
- https://bitbucket.org/fenics-project/dolfin/pull-requests/565
- https://bitbucket.org/fenics-project/dolfin/pull-requests/566
- https://bitbucket.org/fenics-project/dolfin/pull-requests/567
that fixes some edge cases of DOLFIN with MeshView (which is used in SMART).
These changes has to be merged into DOLFIN before we can make a release of DOLFIN (and then push a new release on conda).
