+++ smart/mesh.py
@@ -431,7 +431,7 @@ class ChildMesh(_Mesh):
         self.intersection_map_parent[mesh_id_set].array()[parent_indices] = 1
 
     def get_intersection_submesh(self, mesh_id_set: FrozenSet[int]):
-        self.intersection_submesh[mesh_id_set] = d.MeshView.create(
+        self.intersection_submesh[mesh_id_set] = d.create_meshview(
             self.intersection_map_parent[mesh_id_set], 1
         )
         self.intersection_submesh[mesh_id_set].init()
@@ -449,4 +449,4 @@ class ChildMesh(_Mesh):
 
     def extract_submesh(self):
         mf_type = "cells" if self.is_volume else "facets"
-        self.dolfin_mesh = d.MeshView.create(self.parent_mesh.mf[mf_type], self.primary_marker)
+        self.dolfin_mesh = d.create_meshview(self.parent_mesh.mf[mf_type], self.primary_marker)
