### 2023-07-20 #5580 @emmetfrancis
Hello all, just wanted to provide a quick update. We are working on integrating these fixes with MeshView (mentioned by Jorgen above) into a new release of SMART, which will allow us to include a performance test demonstrating use of parallelization. Once we complete this (looks like next week at this point), this should resolve all of the raised issues. I will update here as soon as we have this new release, really appreciate all of the feedback and suggestions!
