### 2023-07-24 db261ab
Only run gmsh on one process
