+++ smart/mesh_tools.py
@@ -175,62 +175,7 @@ def create_ellipsoids(
     Returns:
         Tuple (mesh, facet_marker, cell_marker)
     """
-    import gmsh
 
-    if np.any(np.isclose(outerRad, 0)):
-        ValueError("One of the outer radii is equal to zero")
-    if np.isclose(hEdge, 0):
-        hEdge = 0.1 * max(outerRad)
-    if np.isclose(hInnerEdge, 0):
-        hInnerEdge = 0.2 * max(outerRad) if np.any(np.isclose(innerRad, 0)) else 0.2 * max(innerRad)
-    if innerRad[0] > outerRad[0] or innerRad[1] > outerRad[1] or innerRad[2] > outerRad[2]:
-        ValueError("Inner ellipsoid does not fit inside outer ellipsoid")
-    # Create the two ellipsoid mesh using gmsh
-    gmsh.initialize()
-    gmsh.option.setNumber("General.Terminal", int(verbose))
-
-    gmsh.model.add("twoellipsoids")
-    # first add ellipsoid 1 of radius outerRad and center (0,0,0)
-    outer_ellipsoid = gmsh.model.occ.addSphere(0, 0, 0, 1.0)
-    gmsh.model.occ.dilate([(3, outer_ellipsoid)], 0, 0, 0, outerRad[0], outerRad[1], outerRad[2])
-    if np.any(np.isclose(innerRad, 0)):
-        # Use outer_ellipsoid only
-        gmsh.model.occ.synchronize()
-        gmsh.model.add_physical_group(3, [outer_ellipsoid], tag=outer_vol_tag)
-        facets = gmsh.model.getBoundary([(3, outer_ellipsoid)])
-        assert len(facets) == 1
-        gmsh.model.add_physical_group(2, [facets[0][1]], tag=outer_marker)
-    else:
-        # Add inner_ellipsoid (radius innerRad, center (0,0,0))
-        inner_ellipsoid = gmsh.model.occ.addSphere(0, 0, 0, 1.0)
-        gmsh.model.occ.dilate(
-            [(3, inner_ellipsoid)], 0, 0, 0, innerRad[0], innerRad[1], innerRad[2]
-        )
-        # Create interface between ellipsoids
-        two_ellipsoids, (outer_ellipsoid_map, inner_ellipsoid_map) = gmsh.model.occ.fragment(
-            [(3, outer_ellipsoid)], [(3, inner_ellipsoid)]
-        )
-        gmsh.model.occ.synchronize()
-
-        # Get the outer boundary
-        outer_shell = gmsh.model.getBoundary(two_ellipsoids, oriented=False)
-        assert len(outer_shell) == 1
-        # Get the inner boundary
-        inner_shell = gmsh.model.getBoundary(inner_ellipsoid_map, oriented=False)
-        assert len(inner_shell) == 1
-        # Add physical markers for facets
-        gmsh.model.add_physical_group(outer_shell[0][0], [outer_shell[0][1]], tag=outer_marker)
-        gmsh.model.add_physical_group(inner_shell[0][0], [inner_shell[0][1]], tag=interface_marker)
-
-        # Physical markers for
-        all_volumes = [tag[1] for tag in outer_ellipsoid_map]
-        inner_volume = [tag[1] for tag in inner_ellipsoid_map]
-        outer_volume = []
-        for vol in all_volumes:
-            if vol not in inner_volume:
-                outer_volume.append(vol)
-        gmsh.model.add_physical_group(3, outer_volume, tag=outer_vol_tag)
-        gmsh.model.add_physical_group(3, inner_volume, tag=inner_vol_tag)
 
     def meshSizeCallback(dim, tag, x, y, z, lc):
         # mesh length is hEdge at the PM (defaults to 0.1*outerRad,
[…]
