### 2023-07-27 648a081
Update README.md to include all example cases and add note about parallelization feature
