### 2023-07-28 #5580 @emmetfrancis
Thank you for your patience while we have assessed some issues with parallelization. At this point, we have found that the problem stems from an issue using `MeshView`'s in parallel in DOLFIN. This bug is not directly related to SMART, but is rather an issue that needs to be addressed in the underlying DOLFIN code. @jorgensd is working on addressing that in DOLFIN. In the meantime, we have added an explanation of the current status of MPI parallelization in SMART within our [README file](https://github.com/RangamaniLabUCSD/smart#readme), under "Functionality Documentation":

> The current version of SMART is not compatible with MPI-based mesh parallelization; this feature is in development pending a future release of DOLFIN addressing some issues when using MeshViews in parallel. However, SMART users can utilize MPI to run multiple simulations in parallel (one mesh per process), as demonstrated in [Example 3 with MPI](https://github.com/RangamaniLabUCSD/smart/blob/development/examples/example3/example3_multimeshMPI.py).

We are hoping we will be able to proceed with this JOSS submission, as the remaining issue depends on DOLFIN and not SMART code. @Kevin-Mattheus-Moerman, if you can let us know your thoughts on the best way to move forward, it would be much appreciated.
