### 2023-09-04 #5580 @Kevin-Mattheus-Moerman
@emmetfrancis @jorgensd can you help provide a summary of what issues are still open. Do I understand it correctly that "performance testing" was recommended, but you are at present unable to add that? How far are you from adding that?

@oalii, @vincent-noel, @mbarzegary are you happy with the changes made and the response to your feedback? Some of you have several checkboxes unticked, could you respond with the key points that you feel still need work?
