### 2023-09-04 #5580 @meg-simula
@Kevin-Mattheus-Moerman SMART relies on the larger finite element software FEniCS, in particular its parallel support. There are missing features in FEniCS that makes part of SMART lack native MPI support. Thus, that performance issue is orthogonal to SMART. The fix in FEniCS is unclear, and definitely not something we should wait on. Hope this clarifies. Also see Emmets comments above.
