+++ .github/workflows/build_docs.yml
@@ -27,7 +27,7 @@ jobs:
 
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
 
       - name: "Install pyvista dependencies"
         run: apt-get update && apt-get install -y libgl1-mesa-glx libxrender1 xvfb nodejs
