### 2023-09-05 #5580 @vincent-noel
I'm happy with the results, and with the answer to my comment about parallelization. The authors did their best, now the problem lies with FEniCS, who have been notified of the issue.
