### 2023-09-05 #5580 @jorgensd
> @emmetfrancis @jorgensd can you help provide a summary of what issues are still open. Do I understand it correctly that "performance testing" was recommended, but you are at present unable to add that? How far are you from adding that?
>
>
>
> @oalii, @vincent-noel, @mbarzegary are you happy with the changes made and the response to your feedback? Some of you have several checkboxes unticked, could you respond with the key points that you feel still need work?

Hello @Kevin-Mattheus-Moerman, thanks for following up. The only remaining issue associated with JOSS review is the "Reproducibility and Functionality" issue from @mbarzegary. We haven't implemented any extra performance tests because of these issues with parallelization, but if there are any specific suggestions from the reviewers on a different performance test they might want to see, we are happy to consider it.
