### 2023-09-06 #5580 @mbarzegary
Hi all,
I believe all the issues are resolved except the high-performance claim. What if the authors remove the term "high-performance" from the paper and PyPI repo? Due to certain issues in the underlying FE package, it's not possible to test the parallelism and mesh partitioning, so we cannot call the software "high-performance" I think.
