### 2023-09-11 #5580 @Kevin-Mattheus-Moerman
Thanks @emmetfrancis and all for the clarifying comments.

@emmetfrancis do you think you can address @mbarzegary 's point on rewording "high performance". I leave it up to you how you address it. I feel you may not need to remove the terms if you want to keep them. I would however comment on the current issue to alert readers/users that some performance aspects cannot at present be evaluated due to this issue with your dependency. So perhaps you can add wording in the paper and in your README/docs about this.
