### 2023-09-12 #5580 @oalii
@Kevin-Mattheus-Moerman  and all the authors,

Please, excuse me again for my delay on this review, not the best of time... I finally managed to get back to it and checked all the boxes of my checklist.

Once again, I would like to acknowledge the solid work done on this library ! I enjoyed reviewing it. As soon as it will be published I will strongly suggest to all the students and young engineers within our team to take a good look at it: It is certainly a example of what a scientific computing library should look like these days.

PS : Exploring the code and examples, some (very) minor comments/questions popped in, I put them below.
