### 2023-09-12 #5580 @oalii
General questions/remarks:

* The precise and explicit definition of units when setting the simulations is very interesting.  In many cases, the main quantities of interest have the same units (e.g. time in sec, concentrations in molecules/um^3...). You did not set any default values for these units. Is this a design decision, so that the user must make the consious choice of selecting an appropriate set of units for its problem? Don't you think that adding default values could spare some times and ease prototyping ?
* Can your model encompass vector and/or tensor fields ?
  * For instance, if I want to model the dynamics of a transmembrane protein sensitive to the curvature of the membrane: I might what to depict the membrane curvature through its extrinsic curvature tensor, Can I add an such ingredient into your simulation framework ?
  * Another example: If I am interested in cytoskeleton dynamics, I might need to describe it with a nematic field or at least a vector field accounting for its polarization. Can I use `VectorElements` to do that?
* You mention somewhere (couldn't retrieve where...) That some notebooks might take some time to run (examples 3, 4 and 6 took some time for me): It could be informative for the user to mention this in the introduction of the corresponding notebooks.
* A naive question (more for my general knowledge than anything else...) about your FEniCS/dolfin dependency: you rely on the "legacy" version of `dolfin`, not the latest `dolfinx` one. If ever, one day, one would want to use your library with FEniCSx/dolfinx, how hard would it be the update?
  * Subsidiary question: Why did you chose to rely on the legacy version and not the current "x" version of these frameworks/libraries ?

Some minor comments/remarks/questions on the example notebooks:

example 1:

* I don't really understand the final regression test:
  * why only on species B ?
  * Where does the `L2norm_saved` come from ?
  * What are you assessing ? seems like the sum ofthe  squared values of the field B at every DOF. Is that it ?

example 2:

* Could be informative to have the law of mass action or/and the differential equation depicting the chemical reaction at stake.

example 3:

* Could be information to have the differential system to solve.
* typo right above cell #14: `Now we repeat this process for other values of cell radii, to test the dependence of steady state calcium concentration on cell size`  → I think you are refearing to the `Aphos` concentration, not calcium concentration, no ?
* Concerning the comparison with analytical expectation, I assume you are using Eq.9 from the reference paper Meyer *et al* (2006) ? It could be informative to explicitly write this equation above cell #18.

example 4:

* Differential system to solve (eqs. 1 to 4 in initial paper, only eq.3 is given) could again be informative in the initial description of the problem in the first cell.

* Assertion error on the last cell:

```
AssertionError                            Traceback (most recent call last)
Cell In[11], line 7
      5 auc_compare = 0.35133404882191754
      6 percent_error = 100*np.abs(auc_cur - auc_compare)/auc_compare
----> 7 assert percent_error < .01,\
      8     f"Failed regression test: Example 4 results deviate {percent_error:.3f}% from the previous numerical solution"

AssertionError: Failed regression test: Example 4 results deviate 0.012% from the previous numerical solution
```

Example 5:

* Warning generated in cell #10:

  ```
  2023-09-12 09:50:29,161 smart.model_assembly - WARNING - Warning! Pre-integrating parameter j1pulse. Make sure that expressions j1pulse appears in have no other time-dependent variables. (model_assembly.py:660)

  2023-09-12 09:50:29,162 smart.model_assembly - WARNING - Warning! Pre-integrating parameter j1pulse. 	Make sure that expressions j1pulse appears in have no other time-dependent variables. 	(model_assembly.py:660)
  ```

Example 6:

* Same warning as in example 5 in cell #8
* Warnings on cell #17:

```
ROOT -2023-09-12 09:56:35,346 FFC - WARNING -   WARNING: The number of integration points for each cell will be: 144 (log.py:153)
ROOT -2023-09-12 09:56:35,346 FFC - WARNING -            Consider using the option 'quadrature_degree' to reduce the number of points (log.py:153)
ROOT -2023-09-12 09:56:35,710 FFC - WARNING -   WARNING: The number of integration points for each cell will be: 144 (log.py:153)
ROOT -2023-09-12 09:56:35,710 FFC - WARNING -            Consider using the option 'quadrature_degree' to reduce the number of points (log.py:153)
```
