### 2023-09-12 #5580 @emmetfrancis
Thank you @oalii for the detailed feedback, much appreciated! I will certainly plan to go through and address/fix the issues in the examples. Regarding your first main questions:

- While setting default units would perhaps somewhat improve usability, it immediately runs the risk of the user not correctly simulating some given problem. We styled this aspect of our software in a similar way to VCell (a popular systems biology software). Note that units become especially confusing (and a source of unintentional mistakes) when coupling 2D-3D problems, as we do here. For instance, 3D concentrations are often specified in some version of molarity (moles per liter), whereas surface concentrations are molecules per area. Reactions between surface and volume species must have rates with carefully chosen units to preserve the integrity of the problem.

- FEniCS can definitely be used to solve vector-valued problems, but we do not have any built-in support for such problems in SMART right now. A workaround for nematic fields could just be writing the vector components as separate species (cyto_x, cyto_y, cyto_z), but I imagine the associated reactions may be messy to write out. We appreciate the suggestion and will certainly consider supporting vector-valued species in future iterations of SMART.
Regarding the curvature sensitive protein, I think this could be captured in our current framework provided one computes the extrinsic curvature over the mesh. In the current version, this might need to be defined as a species that remains constant over the surface and can then be included in the reaction rate for the protein of interest. In future iterations, we hope to include parameters that can vary over the domain.

- There are associated notes in the intros to Examples 3 and 6 saying it may take an hour or longer to execute, I can add a similar note for Example 4.

- It would take quite a bit of rewriting to build this software off of dolfinx rather than legacy dolfin. We developed this in legacy dolfin where the tools for mixed dimensional systems are more mature (see [this paper](https://dl.acm.org/doi/fullHtml/10.1145/3471138)). This may be a future direction we go as FEniCSx continues to mature, but currently the transition is not planned for SMART.
