### 2023-09-12 #5580 @emmetfrancis
> @emmetfrancis could you clarify if the above issue with FENiCs could be circumvented by using an older version? If so we could perhaps use that on this review.

There is no version of FEniCS that would support parallelization for our software currently. @jorgensd can provide a more detailed explanation, as he has been the one working on fixing this parallelization issue in the FEniCS source code.
