### 2023-09-12 #5580 @jorgensd
> > @emmetfrancis could you clarify if the above issue with FENiCs could be circumvented by using an older version? If so we could perhaps use that on this review.
>
> There is no version of FEniCS that would support parallelization for our software currently. @jorgensd can provide a more detailed explanation, as he has been the one working on fixing this parallelization issue in the FEniCS source code.

The issue relates to a corner case in the `MeshView` implementation from https://dl.acm.org/doi/abs/10.1145/3471138
which is used under the hood of SMART to solve PDEs on sub-domains and interfaces (and represent quantities on these).
To the best of my knowledge, the following corner case has never worked as it should with MeshView:
A process having a vertex, but not having an entity of the MeshView in question (can happen if an interface aligns with a process boundary).
Trying to resolve this, and other parallel issues popping up during benchmarking with SMART, I made several fixes to the core code in DOLFIN, Ref:
https://bitbucket.org/fenics-project/dolfin/pull-requests/563
https://bitbucket.org/fenics-project/dolfin/pull-requests/565
https://bitbucket.org/fenics-project/dolfin/pull-requests/567
I've not found a nice way of fixing the corner case mentioned above yet.
The corner case mentioned above does not happen on every parallel run; it depends on the number of process used for any current grid.
