### 2023-09-12 #5580 @jorgensd
> A naive question (more for my general knowledge than anything else...) about your FEniCS/dolfin dependency: you rely on the "legacy" version of `dolfin`, not the latest `dolfinx` one. If ever, one day, one would want to use your library with FEniCSx/dolfinx, how hard would it be the update?
>
>     * Subsidiary question: Why did you chose to rely on the legacy version and not the current "x" version of these frameworks/libraries ?

To add even more context to this. The development of this library started back in 2019. DOLFINx reached a state the core developers thought was mature in 2022, which is when it was recommend to start new projects in DOLFINx: https://fenicsproject.discourse.group/t/the-new-dolfinx-solver-is-now-recommended-over-dolfin/8249
As @emmetfrancis  stated, the current work rely on the MeshView works of Daversin et al, and this is currently in development in DOLFINx in: https://github.com/FEniCS/dolfinx/pull/2271 but has at the time of writing not gotten merged into the main part of the code.
