### 2023-09-14 #5580 @emmetfrancis
Hello @oalii, thanks again for the comments. I have been through the examples and just need to make some tweaks/clarifications where you have suggested. However, the assertion error should definitely not occur in Example 4 (I do not get that at least). Can you clarify which version of SMART you installed to help me with troubleshooting this?
