### 2023-09-15 #5580 @oalii
@emmetfrancis ,
Thanks for addressing this issue. I ran the example 4 notebook in the docker image `smart-lab:lastest`:
```
REPOSITORY                           TAG       IMAGE ID       CREATED        SIZE
ghcr.io/rangamanilabucsd/smart-lab   latest    e42e3f9dfe3a   2 months ago   4.33GB
```
FYI: I am running this on a MacBookPro M1Pro. Could it be a hardware-related issue ?
