### 2023-09-15 #5580 @emmetfrancis
Thanks again, @oalii. It seems likely that the example 4 discrepancy is indeed hardware-related (@finsberg tested on Mac and saw the same thing). We updated the tolerance on the regression test so it should pass for Mac now.

Here are some responses to the smaller concerns with the examples. All are fixed with [this PR ](https://github.com/RangamaniLabUCSD/smart/pull/138) to SMART. Please let me know if there are any remaining concerns

example 1:

- The final regression test is performed by comparing the L2 norm of species B over the domain to a previously stored value (L2norm_saved). I added some text in the notebook to clarify this. Note that we could have chosen a norm of species A, we just need some scalar quantity to compare with previous output.

example 2:

- I added the differential equations for easy reference.

example 3:

- I added the differential equation for easy reference, fixed the typo and added Eq 9 from Meyer et al for easy reference.

example 4:

- I added the differential equation for easy reference.
- As mentioned above, the assertion error seems to be hardware specific. We set the tolerance higher so that this example passes for Mac.

Example 5:

- Thank you for pointing out this warning; it is certainly confusing and I do not think should be user-facing. The associated concern is actually not applicable to the current version of SMART so I have removed this warning and added some associated details in the doc strings for reference in future development.

Example 6:

- Thank you for noting these warnings. They should no longer appear now, I specified a quadrature degree for use in FEniCS (4th degree to be safe).
