### 2023-09-15 143a484
Remove high performance claim from JOSS paper
Remove high performance claim from JOSS paper
