### 2023-09-15 #5580 @emmetfrancis
Regarding the high-performance claim, we have removed "high-performance" from the text of the JOSS paper in [this PR](https://github.com/RangamaniLabUCSD/smart/pull/139). Note that we also give some extra context regarding the state of parallelization in SMART within our README:

> The current version of SMART is not compatible with MPI-based mesh parallelization; this feature is in development pending a future release of DOLFIN addressing some issues when using MeshViews in parallel. However, SMART users can utilize MPI to run multiple simulations in parallel (one mesh per process), as demonstrated in [Example 3 with MPI](https://github.com/RangamaniLabUCSD/smart/blob/development/examples/example3/example3_multimeshMPI.py).
