### 2023-09-15 4a6a626
Also remove "high-performance" from the README
