### 2023-09-18 justinlaughlin/stubs#119 @emmetfrancis
Performance testing will be revisited upon future development of parallelization features
