### 2023-10-04 #5580 @Kevin-Mattheus-Moerman
@oalii @mbarzegary I see all boxes are ticked and the authors have implemented changes in the wording around performance. Can you please confirm that you are happy to recommend acceptance in JOSS (or alternatively summarise any remaining issues). Thanks!
