### 2023-10-15 #5580 @Kevin-Mattheus-Moerman
@emmetfrancis given the favourable reviews above, we will now work towards processing this for acceptance in JOSS. I have just created the final checklist. Can you work on the author's points ☝️
