### 2023-10-17 fad024b
switch to sphinx-autoapi2
