### 2023-10-17 7d9a470
specify version of jupyter book
