### 2023-10-17 1681bf6
return to original (let pip decide versions) - fails build docs
