### 2023-10-18 e89a35d
Add guards for importing ufl_legacy
