### 2023-10-18 b0146a8
Upgrade docker image
