+++ .github/workflows/run_examples_mpi.yml
@@ -17,7 +17,7 @@ jobs:
   run-examples:
     runs-on: ubuntu-22.04
     # Runs against FEniCS main branch built the 16-01-2023
-    container: ghcr.io/scientificcomputing/fenics-gmsh:2023-04-21
+    container: ghcr.io/scientificcomputing/fenics-gmsh:2023-08-16
 
     env:
       PYVISTA_OFF_SCREEN: true
