+++ .github/workflows/build_docs.yml
@@ -15,7 +15,7 @@ jobs:
 
   build:
     runs-on: ubuntu-22.04
-    container: ghcr.io/scientificcomputing/fenics-gmsh:2023-04-21
+    container: ghcr.io/scientificcomputing/fenics-gmsh:2023-08-16
     env:
       DEB_PYTHON_INSTALL_LAYOUT: deb_system
       PUBLISH_DIR: ./_build/html
