+++ .github/workflows/test_fenics_smart.yml
@@ -17,7 +17,7 @@ jobs:
   check-code:
     runs-on: ubuntu-22.04
     # Runs against FEniCS main branch built the 16-01-2023
-    container: ghcr.io/scientificcomputing/fenics-gmsh:2023-04-21
+    container: ghcr.io/scientificcomputing/fenics-gmsh:2023-08-16
     steps:
       # This action sets the current path to the root of your github repo
       - uses: actions/checkout@v3
