### 2023-10-18 #5580 @emmetfrancis
The latest version is v2.1.6 and has been archived on Zenodo. DOI is [10.5281/zenodo.10019502](https://doi.org/10.5281/zenodo.10019502). I have double checked all items on the checklist so we should be good to move forward from our end.
