### 2023-10-18 4b8aacd
Spell out USA
