### 2023-10-18 #5580 @emmetfrancis
Oh, thank you for catching these items, my apologies. I have fixed them and made a new release associated with the affiliations change in the JOSS paper, 2.1.7. The DOI for 2.1.7 is https://doi.org/10.5281/zenodo.10019519, and the DOI for all versions of SMART is https://zenodo.org/doi/10.5281/zenodo.10019463.
