### 2023-10-19 #5580 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
